+++ spec/rubocop/fixtures/app/channels/bad_channel.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class TestChannel < ApplicationCable::Channel
+  def subscribed
+    @bad_var = 'bad'
+  end
+end