+++ spec/rubocop/remote_disconnect_spec.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe Anycable::Compatibility::Anycable::RemoteDisconnect do
+  include_context "cop spec"
+
+  it 'registers offense for remote disconnection attempt' do
+    inspect_source(<<-RUBY.strip_indent)
+      class MyChannel < ApplicationCable::Channel
+        def subscribed
+          ActionCable.server.remote_connections.where(current_user: user).disconnect
+        end
+      end
+    RUBY
+
+    expect(cop.offenses.size).to be(1)
+    expect(cop.messages.first).to eq("Disconnecting remote clients is not supported in AnyCable")
+  end
+end